requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aioredis
pydantic
paho-mqtt
python-dotenv
redis